+++ tracker/tracker/src/main/modules/Network/fetchProxy.ts
@@ -81,8 +81,10 @@ export class ResponseProxyHandler<T extends Response> implements ProxyHandler<T>
               // @ts-ignore
               readerReceivedValue = new Uint8Array(result.value)
             } else {
+              // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
               const newValue = new Uint8Array(readerReceivedValue.length + result.value!.length)
               newValue.set(readerReceivedValue)
+              // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
               newValue.set(result.value!, readerReceivedValue.length)
               readerReceivedValue = newValue
             }
@@ -125,7 +127,7 @@ export class FetchProxyHandler<T extends typeof fetch> implements ProxyHandler<T
     private readonly isServiceUrl: (url: string) => boolean,
   ) {}
 
-  public apply(target: T, thisArg: typeof window, argsList: [RequestInfo | URL, RequestInit]) {
+  public apply(target: T, _: typeof window, argsList: [RequestInfo | URL, RequestInit]) {
     const input = argsList[0]
     const init = argsList[1]
 
@@ -141,6 +143,15 @@ export class FetchProxyHandler<T extends typeof fetch> implements ProxyHandler<T
     const item = new NetworkMessage(this.ignoredHeaders, this.setSessionTokenHeader, this.sanitize)
     this.beforeFetch(item, input as RequestInfo, init)
 
+    this.setSessionTokenHeader((name, value) => {
+      argsList[1] = {
+        ...argsList[1],
+        headers: {
+          ...argsList[1]?.headers,
+          [name]: value,
+        },
+      }
+    })
     return (<ReturnType<T>>target.apply(window, argsList))
       .then(this.afterFetch(item))
       .catch((e) => {